@@ -1,13 +1,18 @@
 {
-  "name": "@baudevs-monorepo/source",
+  "name": "@baudevs/monorepo",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {},
   "private": true,
+  "workspaces": [
+    "apps/*",
+    "libs/*"
+  ],
+  "packageManager": "bun@1.1.1",
   "dependencies": {
-    "next": "14.2.16",
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "next": "15.1.2",
+    "react": "19.0.0",
+    "react-dom": "19.0.0"
   },
   "keywords": [
     "baudevs",
@@ -38,35 +43,37 @@
     "@nx/playwright": "20.2.2",
     "@nx/workspace": "20.2.2",
     "@playwright/test": "^1.36.0",
-    "@swc-node/register": "~1.9.1",
-    "@swc/core": "~1.5.7",
+    "@swc-node/register": "^1.10.9",
+    "@swc/cli": "^0.5.2",
+    "@swc/core": "^1.10.1",
     "@swc/helpers": "~0.5.11",
-    "@testing-library/react": "15.0.6",
+    "@swc/jest": "~0.2.36",
+    "@testing-library/react": "^16.1.0",
     "@types/jest": "^29.5.12",
-    "@types/node": "18.16.9",
-    "@types/react": "18.3.1",
-    "@types/react-dom": "18.3.0",
-    "autoprefixer": "10.4.13",
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
+    "autoprefixer": "^10.4.20",
     "babel-jest": "^29.7.0",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "14.2.16",
+    "eslint": "^9.10.0",
+    "eslint-config-next": "^15.1.2",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jsx-a11y": "6.10.1",
-    "eslint-plugin-playwright": "^1.6.2",
-    "eslint-plugin-react": "7.35.0",
-    "eslint-plugin-react-hooks": "5.0.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-playwright": "^2.1.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.1.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-environment-node": "^29.7.0",
     "nx": "20.2.2",
-    "postcss": "8.4.38",
-    "prettier": "^2.6.2",
-    "tailwindcss": "3.4.3",
+    "postcss": "^8.4.49",
+    "prettier": "^3.4.2",
+    "tailwindcss": "^3.4.17",
     "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
+    "ts-node": "^10.9.2",
     "tslib": "^2.3.0",
-    "typescript": "~5.6.2",
-    "typescript-eslint": "^8.13.0"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.1"
   }
 }
