@@ -1,15 +1,15 @@
-const { FlatCompat } = require('@eslint/eslintrc');
-const js = require('@eslint/js');
-const { fixupConfigRules } = require('@eslint/compat');
-const nx = require('@nx/eslint-plugin');
-const baseConfig = require('../../eslint.config.js');
+import { FlatCompat } from '@eslint/eslintrc';
+import js from '@eslint/js';
+import { fixupConfigRules } from '@eslint/compat';
+import nx from '@nx/eslint-plugin';
+import baseConfig from '../../eslint.config.mjs';
 
 const compat = new FlatCompat({
   baseDirectory: __dirname,
   recommendedConfig: js.configs.recommended,
 });
 
-module.exports = [
+export default [
   ...fixupConfigRules(compat.extends('next')),
 
   ...fixupConfigRules(compat.extends('next/core-web-vitals')),
