@@ -1,6 +1,6 @@
-const baseConfig = require('../../eslint.config.js');
+import baseConfig from '../../eslint.config.mjs';
 
-module.exports = [
+export default [
   ...baseConfig,
   {
     files: ['**/*.json'],
